Phase 4 Wave A: move router.ts into core/
router.ts is the router-only piece of core/ (router + data layer +
persistence + typed API client, per the north-star plan) — a
mechanical relocation ahead of the larger API-client design work,
which is scoped separately.

Updates all 30 importers across apps/web (root files, mp/, surfaces/*)
plus stale doc-comment references in apps/api/src/platform/spa-paths.ts
and apps/e2e/src/walkthrough.test.ts. Baselines resealed for the new
path.
