+++ apps/web/src/App.catalog.test.ts
@@ -5,7 +5,7 @@ import { createRoot } from 'react-dom/client';
 import { afterEach, describe, expect, it, vi } from 'vitest';
 import { App } from './App.js';
 import { AuthProvider } from './AuthContext.js';
-import { NAVIGATE_EVENT, type NavigateEventDetail } from './router.js';
+import { NAVIGATE_EVENT, type NavigateEventDetail } from './core/router.js';
 import i18n from './i18n/index.js';
 
 async function flushEffects() {
