+++ apps/web/src/surfaces/admin/AdminConsole.tsx
@@ -9,7 +9,7 @@ import { SuggestionsPanel } from './SuggestionsPanel.js';
 import { WaitlistPanel } from './WaitlistPanel.js';
 import { AdminAssessmentsPanel } from './AdminAssessmentsPanel.js';
 import { fetchAdminSummary, type AdminSummary, type OperatorAlert } from './adminApi.js';
-import { ADMIN_SECTIONS, adminPath, type AdminSection } from '../../router.js';
+import { ADMIN_SECTIONS, adminPath, type AdminSection } from '../../core/router.js';
 
 /**
  * One place for everything an operator can do.
