+++ apps/web/src/surfaces/studio/StudioWelcomeView.tsx
@@ -3,7 +3,7 @@ import { createPortal } from 'react-dom';
 import { useTranslation } from 'react-i18next';
 import { InteractiveMascot, type MascotEmotion } from '../../Mascot.js';
 import { PixelIcon } from '../../PixelIcon.js';
-import { studioPath } from '../../router.js';
+import { studioPath } from '../../core/router.js';
 import { BuildProgressChecklist } from '../../BuildProgressChecklist.js';
 import { isStudioOnboarded, markStudioOnboarded, resolveWelcomeToken } from './studioWelcome.js';
 import { pollDelayMs } from './studioStatusPoll.js';
