+++ eslint-rules/comment-prose-baseline.json
@@ -275,7 +275,7 @@
     "apps/api/src/platform/rate-limit.ts": 59,
     "apps/api/src/platform/server.ts": 59,
     "apps/api/src/platform/spa-paths.test.ts": 101,
-    "apps/api/src/platform/spa-paths.ts": 437,
+    "apps/api/src/platform/spa-paths.ts": 440,
     "apps/api/src/platform/store.test.ts": 394,
     "apps/api/src/platform/store.ts": 625,
     "apps/api/src/platform/submission-status.test.ts": 35,
@@ -355,7 +355,7 @@
     "apps/e2e/src/globalSetup.ts": 153,
     "apps/e2e/src/resilience.test.ts": 356,
     "apps/e2e/src/studio-shell.test.ts": 877,
-    "apps/e2e/src/walkthrough.test.ts": 713,
+    "apps/e2e/src/walkthrough.test.ts": 714,
     "apps/e2e/vitest.config.ts": 93,
     "apps/web/src/AccountDeletionControl.tsx": 40,
     "apps/web/src/AccountSettingsModal.tsx": 12,
@@ -388,6 +388,8 @@
     "apps/web/src/connectCollapse.ts": 26,
     "apps/web/src/ContactPage.tsx": 64,
     "apps/web/src/ContributionsSetting.tsx": 197,
+    "apps/web/src/core/router.test.ts": 451,
+    "apps/web/src/core/router.ts": 2406,
     "apps/web/src/creatorGames.test.ts": 13,
     "apps/web/src/creatorGames.ts": 115,
     "apps/web/src/CreatorMetricsPanel.test.tsx": 108,
@@ -499,8 +501,6 @@
     "apps/web/src/RemixPanel.tsx": 3870,
     "apps/web/src/ReportGameButton.test.tsx": 137,
     "apps/web/src/ReportGameButton.tsx": 180,
-    "apps/web/src/router.test.ts": 451,
-    "apps/web/src/router.ts": 2406,
     "apps/web/src/ScorecardPanel.test.tsx": 185,
     "apps/web/src/ScorecardPanel.tsx": 388,
     "apps/web/src/selfBuildCopy.test.ts": 27,
