+++ eslint-rules/module-size-baseline.json
@@ -530,6 +530,8 @@
     "apps/web/src/ContactPage.test.tsx": 103,
     "apps/web/src/ContactPage.tsx": 133,
     "apps/web/src/ContributionsSetting.tsx": 105,
+    "apps/web/src/core/router.test.ts": 449,
+    "apps/web/src/core/router.ts": 670,
     "apps/web/src/createInitialPrompt.test.ts": 18,
     "apps/web/src/createInitialPrompt.ts": 4,
     "apps/web/src/CreatePage.test.tsx": 116,
@@ -564,7 +566,7 @@
     "apps/web/src/gamepadSpike.test.ts": 191,
     "apps/web/src/gamepadSpike.ts": 173,
     "apps/web/src/GamePage.test.tsx": 410,
-    "apps/web/src/GamePage.tsx": 396,
+    "apps/web/src/GamePage.tsx": 400,
     "apps/web/src/gamePageApi.ts": 36,
     "apps/web/src/gamePlayer.bridge.test.ts": 219,
     "apps/web/src/gamePlayer.hiddenFrame.test.tsx": 118,
@@ -674,8 +676,6 @@
     "apps/web/src/RemixPanel.tsx": 1796,
     "apps/web/src/ReportGameButton.test.tsx": 96,
     "apps/web/src/ReportGameButton.tsx": 53,
-    "apps/web/src/router.test.ts": 449,
-    "apps/web/src/router.ts": 670,
     "apps/web/src/ScorecardPanel.test.tsx": 165,
     "apps/web/src/ScorecardPanel.tsx": 170,
     "apps/web/src/selfBuildCopy.test.ts": 144,
